+++ core/runtime/plugin_registry.cpp
@@ -5,11 +5,15 @@
 
 #include "plugin_registry.hpp"
 
+#include "log_macros.hpp"
+
 namespace scope::runtime
 {
 
 void PluginRegistry::registerPlugin(PluginInfo info)
 {
+    SCOPE_LOG_INFO("runtime.plugin", "register plugin=" + info.name);
+
     m_plugins.push_back(std::move(info));
 }